change context root of support tool
The support tool is deployed in the same k8 cluster as student, proctor apps.
Setting the default context root to '/supporttool'.

Note that k8 ingress configuration could be used to accomplish the same effect:
annotations:
    ingress.kubernetes.io/add-base-url: "true"
    ingress.kubernetes.io/rewrite-target: /

Except that in the current k8 configuration, the add-base-url annotation did not
rewrite the base-url html tag.