Remove object mapper declared as a bean.
Once an object mapper is declared as a bean, it is used in for mapping for http requests.
tds-common doesn’t declare an object mapper as a bean, it defines a builder customizer to modify the object mapper used in the http message converter.
In support tool, avoid declaring an object mapper as a bean. To not use a different object mapper for http mapping, declare a bean that produces an object mapper instead.