fix link destination when clicking on home page logo
Link used with the logo was an absolute location.  When the app was deployed in k8 environment under a context path, the link was no longer valid.
The k8 ingress used to redirect that location to the support-tool.  The ingress has been updated an the re-direction removed.
Switching to using relative paths via angular routes.